hal: renesas: ra: add portable patch to disable C-CACHE on CPU1/CM33
A known issue exists with C-CACHE version 2 and MRAM on the CM33 core
of dual-core devices. Currently, the MRAM controller does not
guarantee that it will avoid concurrency issues with C-CACHE by
flushing data before writing to ROMABLE space.

This commit adds a patch to disable C-CACHE on the CM33 core of dual
core devices as a temporary solution until the CACHE management driver
for C-CACHE is supported, then C-CACHE maintenance should be added
inside MRAM driver.

Signed-off-by: Khoa Tran <khoa.tran.yj@bp.renesas.com>
Signed-off-by: The Nguyen <the.nguyen.yf@renesas.com>