hal: renesas: ra: correct AGT compare match register set value
Currently, the register setting value for the Period Count is reduced by
one from the application input. This differs from the Compare Match
value, which is set directly by the application input value.

This commit updates the Compare Match Register setting to match the
Period Count setting. It reduces the value by one to maintain consistency
with the Period Count setting and avoid issues when the Compare Value is
greater than the maximum Period Count.

Signed-off-by: Khai Cao <khai.cao.xk@renesas.com>
Signed-off-by: The Nguyen <the.nguyen.yf@renesas.com>